+++ src/App.jsx
@@ -1,22 +1,13 @@
-import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 import { theme } from "./styles/theme";
 import GlobalStyle from "./styles/GlobalStyle";
-import Header from "./components/Header/Header";
-import Home from "./components/Header/Hone";
-import Post from "./components/Header/Post";
+import Router from "./routes/Router";
 
 function App() {
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyle />
-      <BrowserRouter>
-        <Header />
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/post" element={<Post />} />
-        </Routes>
-      </BrowserRouter>
+      <Router />
     </ThemeProvider>
   );
 }
