+++ src/routes/Router.jsx
@@ -1,0 +1,16 @@
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Home from "../pages/Home";
+import Post from "../pages/Post";
+
+function Router() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/post" element={<Post />} />
+      </Routes>
+    </BrowserRouter>
+  );
+}
+
+export default Router;
